greedy algorithm
this one project has featured two of the greedy algorithms

one activity selection algorithm which selects the atmost events to attend without overlapping

and the second one was the huffman coding

This Python code implements Huffman Coding for text compression. It first calculates the frequency of each character in a string, builds a min-heap to construct a binary Huffman tree, and then generates binary codes for each character based on their frequency. Finally, it encodes the input string using the generated Huffman codes.